refactoring to enable customisable context menu for chapters mode, which will result less confusion for users.
