+++ src/main/resources/uk/yermak/audiobookconverter/fx/fxml_converter.fxml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
-
-
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.*?>
 <?import javafx.scene.input.*?>
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 <?import javafx.stage.Screen?>
+<?import uk.yermak.audiobookconverter.fx.BookStructureComponent?>
 <?import uk.yermak.audiobookconverter.fx.FileListComponent?>
 <VBox fx:controller="uk.yermak.audiobookconverter.fx.FilesController"
       xmlns:fx="http://javafx.com/fxml"
@@ -142,8 +141,6 @@
         <Insets top="10" right="10" bottom="10" left="10"/>
     </padding>
 
-    <!--    <HBox VBox.vgrow="ALWAYS" prefHeight="${screen.visualBounds.height * 0.45}">-->
-
         <TabPane fx:id="filesChapters" HBox.hgrow="ALWAYS">
             <padding>
                 <Insets top="10" right="10" bottom="10" left="10"/>
@@ -201,31 +198,6 @@
                             </tooltip>
                         </Button>
 
-<!--
-                        <Button fx:id="upButton" text="Up" minWidth="${screen.visualBounds.width * 0.05}"
-                                onAction="#moveUp"
-                                disable="false">
-                            <tooltip>
-                                <Tooltip text="Move selected file up"/>
-                            </tooltip>
-                        </Button>
-                        <Button fx:id="downButton" text="Down" minWidth="${screen.visualBounds.width * 0.05}"
-                                onAction="#moveDown"
-                                disable="false" GridPane.valignment="TOP">
-                            <tooltip>
-                                <Tooltip text="Move selected file down"/>
-                            </tooltip>
-                        </Button>
--->
-<!--
-                        <Button fx:id="removeButton" text="Remove"
-                                minWidth="${screen.visualBounds.width * 0.05}"
-                                onAction="#removeFiles" disable="false">
-                            <tooltip>
-                                <Tooltip text="Removes selected files from the list"/>
-                            </tooltip>
-                        </Button>
--->
                         <Button fx:id="clearButton" text="Clear" minWidth="${screen.visualBounds.width * 0.05}"
                                 onAction="#clear"
                                 disable="false">
@@ -274,64 +246,6 @@
                             </tooltip>
                         </Button>
 
-<!--
-                        <Button text="Edit" minWidth="${screen.visualBounds.width * 0.05}"
-                                onAction="#edit"
-                                disable="false" GridPane.valignment="TOP">
-                            <font>
-                                <Font name="Arial Black" size="12.0"/>
-                            </font>
-
-                            <tooltip>
-                                <Tooltip text="Edit chapter"/>
-                            </tooltip>
-                        </Button>
--->
-
-<!--
-                        <Button text="Combine" minWidth="${screen.visualBounds.width * 0.05}"
-                                onAction="#combine"
-                                disable="false">
-                            <tooltip>
-                                <Tooltip text="Combine selected"/>
-                            </tooltip>
-                        </Button>
-                        <Button text="Split" minWidth="${screen.visualBounds.width * 0.05}"
-                                onAction="#split"
-                                disable="false" GridPane.valignment="TOP">
-                            <tooltip>
-                                <Tooltip text="Split here"/>
-                            </tooltip>
-                        </Button>
--->
-
-
-<!--
-                        <Button text="Up" minWidth="${screen.visualBounds.width * 0.05}"
-                                onAction="#moveUp"
-                                disable="false">
-                            <tooltip>
-                                <Tooltip text="Move selected file up"/>
-                            </tooltip>
-                        </Button>
-                        <Button text="Down" minWidth="${screen.visualBounds.width * 0.05}"
-                                onAction="#moveDown"
-                                disable="false" GridPane.valignment="TOP">
-                            <tooltip>
-                                <Tooltip text="Move selected file down"/>
-                            </tooltip>
-                        </Button>
--->
-
-<!--
-                        <Button text="Remove"
-                                minWidth="${screen.visualBounds.width * 0.05}"
-                                onAction="#removeFiles" disable="false">
-                            <tooltip>
-                                <Tooltip text="Removes selected files from the list"/>
-                            </tooltip>
-                        </Button>
--->
                         <Button text="Clear" minWidth="${screen.visualBounds.width * 0.05}"
                                 onAction="#clear"
                                 disable="false">
@@ -352,7 +266,7 @@
 
 
                     </HBox>
-                    <TreeTableView fx:id="bookStructure" editable="true" VBox.vgrow="ALWAYS">
+                    <BookStructureComponent fx:id="bookStructure" editable="true" VBox.vgrow="ALWAYS">
                         <columns>
                             <TreeTableColumn text="Title" fx:id="chapterColumn" sortable="false"
                                              prefWidth="${screen.visualBounds.width * 0.1}"/>
@@ -360,50 +274,10 @@
                             <TreeTableColumn text="Details" fx:id="detailsColumn" sortable="false"
                                              prefWidth="${screen.visualBounds.width * 0.2}"/>
                         </columns>
-                    </TreeTableView>
+                    </BookStructureComponent>
                 </VBox>
             </Tab>
         </TabPane>
-
-    <!--<GridPane>
-        <rowConstraints>
-            <RowConstraints vgrow="SOMETIMES"/>
-            <RowConstraints vgrow="ALWAYS"/>
-            <RowConstraints vgrow="NEVER"/>
-        </rowConstraints>-->
-
-    <!--
-                <Accordion expandedPane="$myPane" GridPane.columnIndex="0" GridPane.rowIndex="0">
-                    <panes>
-                        <TitledPane fx:id="myPane" text="Files">
-                            <VBox spacing="5">
-                                <padding>
-                                    <Insets top="10" right="5" bottom="10" left="5"/>
-                                </padding>
-                            </VBox>
-                        </TitledPane>
-                        <TitledPane text="Chapters">
-                            <VBox spacing="5">
-                                <padding>
-                                    <Insets top="10" right="5" bottom="10" left="5"/>
-                                </padding>
-
-
-                            </VBox>
-                        </TitledPane>
-                    </panes>
-                </Accordion>
-    -->
-    <!--
-                <Pane GridPane.columnIndex="0" GridPane.rowIndex="1"/>
-                <Pane GridPane.columnIndex="0" GridPane.rowIndex="2" GridPane.halignment="CENTER">
-                    <padding>
-                        <Insets top="0" right="0" bottom="10" left="0"/>
-                    </padding>
-                </Pane>
-    -->
-    <!--     </GridPane>-->
-    <!--</HBox>-->
 
     <fx:include source="/uk/yermak/audiobookconverter/fx/mediaplayer.fxml"/>
